fix: move the find-build logic for firebase during preprocessing
This allows us to retrigger submission failures. Where, if previously they
could not be either uploaded or found on firebase by Tramline, but were manually
uploaded to firebase; we can hit Retry on Tramline and it will
correctly find it.

Before this change, that was not possible, because we were directly
moving from created --> failed. The retry functionality works by
checking if there's an event that can take us to the
last_stable_status (that was stored before failure) and calls that.

The state 'created' as 'last_stable_status' does not have any events
that take it there as its the initial status, hence the retry
functionality would do nothing in this case.

It's better to do this find-logic after we move to preprocessing
state, so that retry can eventually work (by manually uploading for eg.).