feat: Show build number next to version name for Previous Releases
Re-using the ReleasePresenter#display_build_number method to display the
build number besides the release version.
This is not the final code that should be shipped, since we only need
to show build numbers of builds that were deployed to 100%, but this
should be a good place to start, hopefully.