fix: Make Bitbucket workspace dropdown switcher work
Bitbucket version control configuration requires the user to select
a workspace to select the code repository.
The workspace dropdown is supposed work by re-rendering the configuration
form via Turbo stream. However, the partial for the TURBO_STREAM format
was missing, which would break the interaction, along with raising a
server-side error.
These changes refactor the HTML partial to render the form in both the
HTML and the TURBO_STREAM formats, to make the drop-down UX work.