fix: Pre-select persisted integration config value
The persisted config value, when serialized, could generate a JSON string
that could be different from the one generated by the
view helper for the drop-down options' values, which would make it so that
the persisted config value is not pre-selected by default.
Thus, we explicitly find the correct config and use that as the default,
rather than using the persisted value.